refactor: 슬라이스 내부 로직 import 상대 경로로 수정
