+++ src/entities/text-parsing/lib/helpers/test/get-attendance-badge-id.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest'
 import { ATTENDANCE_BADGE_MOCK_DATA } from '@/entities/attendance-badge/api/mocks/badge-list'
-import { getAttendanceBadgeId } from '@/entities/text-parsing'
+import { getAttendanceBadgeId } from '../parse-text'
 
 describe('출석 통계를 바탕으로 배지를 계산한다.', () => {
   test('배지가 있는 경우 배지 아이디를 반환한다.', () => {
